requirements.txt
torch 
torchvision 
transformers 
pillow 
matplotlib 
seaborn
fpdf